Fix specular light

+++ src/color.c
@@ -6,7 +6,7 @@
 /*   By: llai <llai@student.42london.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/20 14:52:21 by llai              #+#    #+#             */
-/*   Updated: 2024/05/05 00:01:30 by llai             ###   ########.fr       */
+/*   Updated: 2024/05/05 00:18:11 by llai             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ t_color	sub_colors(t_color c1, t_color c2)
 	return (res);
 }
 
-t_color	mul_color(t_color c, double scalar)
+t_color	mul_color(t_color c, float scalar)
 {
 	t_color	res;
